Refactor to avoid compiler crash in Swift 4.2.
The project does not compile on Linux with Swift 4.1 due to
some compiler crash that's only reproduceable on Ubuntu.
That's why I'm trying to get it to compiler on Swift 4.2

Currently there is a compiler bug on 4.2 that can be avoided
with the following refactor. The Jira ticket to track the bug:
https://bugs.swift.org/browse/SR-7858